+++ uui/components/pickers/PickerModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DataRowProps, DataSourceState, IHasCaption, PickerBaseOptions } from '@epam/uui-core';
+import { IHasCaption, PickerBaseOptions, PickerRenderRowParams } from '@epam/uui-core';
 import { IconContainer, PickerModalArrayProps, PickerModalOptions, PickerModalScalarProps, handleDataSourceKeyboard, usePickerModal } from '@epam/uui-components';
 import { DataPickerRow } from './DataPickerRow';
 import { Text } from '../typography';
@@ -10,7 +10,6 @@ import {
 import { LinkButton, Button } from '../buttons';
 import { SearchInput, Switch } from '../inputs';
 import { DataPickerBody } from './DataPickerBody';
-import { PickerItem } from './PickerItem';
 import { i18n } from '../../i18n';
 import { settings } from '../../settings';
 
@@ -27,7 +26,6 @@ export function PickerModal<TItem, TId>(props: PickerModalProps<TItem, TId>) {
         selection,
         dataSourceStateLens,
         dataSourceState,
-        getDataSourceState,
         getName,
         clearSelection,
         getRows,
@@ -37,38 +35,17 @@ export function PickerModal<TItem, TId>(props: PickerModalProps<TItem, TId>) {
         handleDataSourceValueChange,
     } = usePickerModal<TItem, TId>(props);
 
-    const getSubtitle = ({ path }: DataRowProps<TItem, TId>, { search }: DataSourceState) => {
-        if (!search) return;
-
-        return path
-            .map(({ value }) => getName(value))
-            .filter(Boolean)
-            .join(' / ');
-    };
-
-    const renderItem = (item: TItem, rowProps: DataRowProps<TItem, TId>, dsState: DataSourceState) => {
-        const { flattenSearchResults } = view.getConfig();
-        return (
-            <PickerItem
-                title={ getName(item) }
-                size="36"
-                dataSourceState={ dsState }
-                { ...(flattenSearchResults ? { subtitle: getSubtitle(rowProps, dataSourceState) } : {}) }
-                { ...rowProps }
-            />
-        );
-    };
-
-    const renderRow = (rowProps: DataRowProps<TItem, TId>) => {
+    const renderRow = (rowProps: PickerRenderRowParams<TItem, TId>) => {
         return props.renderRow ? (
             props.renderRow(rowProps, dataSourceState)
         ) : (
             <DataPickerRow
                 { ...rowProps }
                 key={ rowProps.rowKey }
-                padding="24"
-                size="36"
-                renderItem={ (item, itemProps) => renderItem(item, itemProps, dataSourceState) }
+                padding={ settings.pickerInput.sizes.body.modalPadding }
+                size={ settings.pickerInput.sizes.body.modalRow }
+                flattenSearchResults={ view.getConfig().flattenSearchResults }
+                getName={ getName }
             />
         );
     };
@@ -110,7 +87,6 @@ export function PickerModal<TItem, TId>(props: PickerModalProps<TItem, TId>) {
     };
 
     const dataRows = getRows();
-    const rows = dataRows.map((row) => renderRow(row));
 
     return (
         <ModalBlocker { ...props }>
@@ -123,7 +99,7 @@ export function PickerModal<TItem, TId>(props: PickerModalProps<TItem, TId>) {
                             onKeyDown={ (e) =>
                                 handleDataSourceKeyboard(
                                     {
-                                        value: getDataSourceState(),
+                                        value: dataSourceState,
                                         onValueChange: handleDataSourceValueChange,
                                         listView: view,
                                         rows: dataRows,
@@ -149,13 +125,15 @@ export function PickerModal<TItem, TId>(props: PickerModalProps<TItem, TId>) {
                 </FlexCell>
                 <DataPickerBody
                     { ...getListProps() }
-                    value={ getDataSourceState() }
+                    value={ dataSourceState }
                     onValueChange={ handleDataSourceValueChange }
-                    search={ dataSourceStateLens.prop('search').toProps() }
                     showSearch={ false }
-                    rows={ rows }
-                    renderNotFound={ renderNotFound }
-                    editMode="modal"
+                    getName={ getName }
+                    rows={ dataRows }
+                    renderRow={ renderRow }
+                    renderEmpty={ renderNotFound }
+                    size={ settings.pickerInput.sizes.body.modalRow }
+                    cx={ css.body }
                 />
                 <ModalFooter padding="24" vPadding="24">
                     {props.renderFooter ? props.renderFooter(getFooterProps()) : renderFooter()}