+++ uui/components/pickers/__tests__/DataPickerBody.test.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { renderHook, renderSnapshotWithContextAsync } from '@epam/uui-test-utils';
 import { ArrayDataSource } from '@epam/uui-core';
 import { DataPickerBody, DataPickerBodyProps } from '../DataPickerBody';
-import { DataPickerRow } from '../DataPickerRow';
 
 const languageLevels = [
     { id: 2, level: 'A1' }, { id: 3, level: 'A1+' }, { id: 4, level: 'A2' }, { id: 5, level: 'A2+' }, { id: 6, level: 'B1' }, { id: 7, level: 'B1+' }, { id: 8, level: 'B2' }, { id: 9, level: 'B2+' }, { id: 10, level: 'C1' }, { id: 11, level: 'C1+' }, { id: 12, level: 'C2' },
@@ -21,13 +20,8 @@ describe('DataPickerBody', () => {
     const requiredProps: DataPickerBodyProps = {
         value: { topIndex: 0 },
         onValueChange: jest.fn(),
-        rows: rows.map(
-            (props) => <DataPickerRow key={ props.id } size="36" renderItem={ (item: string) => <div>{item}</div> } id={ props.id } rowKey={ props.rowKey } index={ props.id } value="" />,
-        ),
-        search: {
-            value: '',
-            onValueChange: jest.fn(),
-        },
+        getName: (item: any) => item.level,
+        rows: rows,
     };
 
     it('should be rendered with minimum props', async () => {
@@ -44,21 +38,10 @@ describe('DataPickerBody', () => {
         const tree = await renderSnapshotWithContextAsync(
             <DataPickerBody
                 { ...requiredProps }
-                editMode="modal"
                 showSearch="auto"
                 maxHeight={ 800 }
                 searchSize="48"
-                rows={ rows.map((props) => (
-                    <DataPickerRow
-                        key={ props.id }
-                        size="36"
-                        renderItem={ (item: string) => <div>{item}</div> }
-                        id={ props.id }
-                        rowKey={ props.rowKey }
-                        index={ props.id }
-                        value={ props.value?.level }
-                    />
-                )) }
+                rows={ rows }
                 onKeyDown={ jest.fn }
                 rowsCount={ 7 }
                 totalCount={ 11 }